Rename enum and class to uppercase
