+++ src/ProgressBadge.ts
